+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [Unreleased]
+
+### Added
+
+* The option to request additional user arguments when starting a process. Can be used for example to set test filters when starting a GTest executable.
+
 ## [0.0.11]
 
 ### Added